Move network connections into logic/network
