+++ packages/sanes-chrome-extension/src/logic/config/index.ts
@@ -1,1 +1,2 @@
-export * from './personafromconfig';
+export * from './codec';
+export * from './configurationfile';
