+++ packages/sanes-chrome-extension/src/logic/network/personafromconfig.ts
