+++ packages/sanes-chrome-extension/src/routes/signup/components/ShowPhraseForm.tsx
@@ -7,7 +7,7 @@ import Switch from 'medulas-react-components/lib/components/Switch';
 import Tooltip from 'medulas-react-components/lib/components/Tooltip';
 import PageLayout from 'medulas-react-components/lib/components/PageLayout';
 import { SIGNUP_ROUTE } from '../../paths';
-import { getPersonaFromConfig } from '../../../logic/config';
+import { getPersonaFromConfig } from '../../../logic/network';
 
 export const SECOND_STEP_SIGNUP_ROUTE = `${SIGNUP_ROUTE}2`;
